+++ packages/designer/src/designer/designer.ts
@@ -67,10 +67,19 @@ export class Designer {
 
     this.project = new Project(this, props.defaultSchema);
 
+    let startTime: any;
+    let src = '';
     this.dragon.onDragstart((e) => {
+      startTime = Date.now() / 1000;
       this.detecting.enable = false;
       const { dragObject } = e;
       if (isDragNodeObject(dragObject)) {
+        const node = dragObject.nodes[0]?.parent;
+        const npm = node?.componentMeta?.npm;
+        src =
+          [npm?.package, npm?.componentName].filter((item) => !!item).join('-') ||
+          node?.componentMeta?.componentName ||
+          '';
         if (dragObject.nodes.length === 1) {
           if (dragObject.nodes[0].parent) {
             // ensure current selecting
@@ -111,6 +120,32 @@ export class Designer {
           if (nodes) {
             loc.document.selection.selectAll(nodes.map((o) => o.id));
             setTimeout(() => this.activeTracker.track(nodes![0]), 10);
+            const endTime: any = Date.now() / 1000;
+            const parent = nodes[0]?.parent;
+            const npm = parent?.componentMeta?.npm;
+            const dest =
+              [npm?.package, npm?.componentName].filter((item) => !!item).join('-') ||
+              parent?.componentMeta?.componentName ||
+              '';
+            this.editor?.emit('designer.drag', {
+              time: (endTime - startTime).toFixed(2),
+              selected: nodes
+                ?.map((n) => {
+                  if (!n) {
+                    return;
+                  }
+                  const npm = n?.componentMeta?.npm;
+                  return (
+                    [npm?.package, npm?.componentName].filter((item) => !!item).join('-') ||
+                    n?.componentMeta?.componentName
+                  );
+                })
+                .join('&'),
+              align: loc?.detail?.near?.align || '',
+              pos: loc?.detail?.near?.pos || '',
+              src,
+              dest,
+            });
           }
         }
       }
@@ -228,7 +263,7 @@ export class Designer {
 
   touchOffsetObserver() {
     this.clearOobxList(true);
-    this.oobxList.forEach(item => item.compute());
+    this.oobxList.forEach((item) => item.compute());
   }
 
   createSettingEntry(editor: IEditor, nodes: Node[]) {
@@ -392,11 +427,16 @@ export class Designer {
   @computed get componentsMap(): { [key: string]: NpmInfo | Component } {
     const maps: any = {};
     this._componentMetasMap.forEach((config, key) => {
-      const view = config.getMetadata().experimental?.view;
-      if (view) {
-        maps[key] = view;
-      } else if (config.npm) {
-        maps[key] = config.npm;
+      const metaData = config.getMetadata();
+      if (metaData.devMode === 'lowcode') {
+        maps[key] = this.currentDocument?.simulator?.createComponent(metaData.schema!);
+      } else {
+        const view = metaData.experimental?.view;
+        if (view) {
+          maps[key] = view;
+        } else if (config.npm) {
+          maps[key] = config.npm;
+        }
       }
     });
     return maps;