+++ packages/designer/src/document/document-model.ts
@@ -66,6 +66,7 @@ export class DocumentModel {
     return this.modalNode || this.rootNode;
   }
 
+  private inited = false;
   constructor(readonly project: Project, schema?: RootSchema) {
     /*
     // TODO
@@ -88,10 +89,11 @@ export class DocumentModel {
     );
 
     this.history = new History(
-      () => this.schema,
+      () => this.export(TransformStage.Serilize),
       (schema) => this.import(schema as RootSchema, true),
     );
     this.setupListenActiveNodes();
+    this.inited = true;
   }
 
   @obx.val private willPurgeSpace: Node[] = [];
@@ -161,6 +163,9 @@ export class DocumentModel {
     }
 
     let node: Node | null = null;
+    if (!this.inited) {
+      schema.id = null;
+    }
     if (schema.id) {
       node = this.getNode(schema.id);
       if (node && node.componentName === schema.componentName) {
@@ -483,7 +488,7 @@ export class DocumentModel {
 
   // add toData
   toData() {
-    const node = this.project?.currentDocument?.export(TransformStage.Serilize);
+    const node = this.project?.currentDocument?.export(TransformStage.Save);
     return { componentsTree: [node] };
   }