+++ packages/designer/src/document/node/transform-stage.ts
@@ -2,5 +2,6 @@ export enum TransformStage {
   Render = 1,
   Serilize = 2,
   Save = 3,
-  Init = 4,
+  Clone = 4,
+  Init = 5,
 }