+++ packages/editor-core/src/hotkey.ts
@@ -1,3 +1,6 @@
+import { globalContext } from './di';
+import { Editor } from './editor';
+
 interface KeyMap {
   [key: number]: string;
 }
@@ -329,6 +332,19 @@ function fireCallback(callback: HotkeyCallback, e: KeyboardEvent, combo?: string
     e.preventDefault();
     e.stopPropagation();
   }
+  const editor = globalContext.get(Editor);
+  const designer = editor.get('designer');
+  const node = designer?.currentSelection?.getNodes()?.[0];
+  const npm = node?.componentMeta?.npm;
+  const selected =
+    [npm?.package, npm?.componentName].filter((item) => !!item).join('-') || node?.componentMeta?.componentName || '';
+  editor?.emit('hotkey.callback.call', {
+    callback,
+    e,
+    combo,
+    sequence,
+    selected,
+  });
 }
 
 export class Hotkey {