+++ packages/editor-skeleton/src/transducers/parse-props.ts
@@ -53,7 +53,7 @@ function propTypeToSetter(propType: PropType): SetterType {
       };
     case 'bool':
       return {
-        componentName: 'NumberSetter',
+        componentName: 'BoolSetter',
         isRequired,
         initialValue: false,
       };