+++ packages/plugin-event-bind-dialog/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.8.16"></a>
+## [0.8.16](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-event-bind-dialog@0.8.15...@ali/lowcode-plugin-event-bind-dialog@0.8.16) (2020-06-15)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-plugin-event-bind-dialog
+
+<a name="0.8.15"></a>
+## [0.8.15](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-event-bind-dialog@0.8.14...@ali/lowcode-plugin-event-bind-dialog@0.8.15) (2020-05-20)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-plugin-event-bind-dialog
+
+<a name="0.8.14"></a>
+## [0.8.14](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-event-bind-dialog@0.8.13...@ali/lowcode-plugin-event-bind-dialog@0.8.14) (2020-05-18)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-plugin-event-bind-dialog
+
+<a name="0.8.13"></a>
+## [0.8.13](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-event-bind-dialog@0.8.12...@ali/lowcode-plugin-event-bind-dialog@0.8.13) (2020-05-15)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-plugin-event-bind-dialog
+
 <a name="0.8.12"></a>
 ## [0.8.12](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-plugin-event-bind-dialog@0.8.11...@ali/lowcode-plugin-event-bind-dialog@0.8.12) (2020-05-13)