+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.8.7"></a>
+## [0.8.7](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-utils@0.8.6...@ali/lowcode-utils@0.8.7) (2020-06-15)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-utils
+
+<a name="0.8.6"></a>
+## [0.8.6](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-utils@0.8.5...@ali/lowcode-utils@0.8.6) (2020-05-18)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-utils
+
+<a name="0.8.5"></a>
+## [0.8.5](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-utils@0.8.4...@ali/lowcode-utils@0.8.5) (2020-05-15)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-utils
+
 <a name="0.8.4"></a>
 ## [0.8.4](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-utils@0.8.3...@ali/lowcode-utils@0.8.4) (2020-05-13)